+++ packages/gatsby-recipes/src/renderer/render.js
@@ -1,6 +1,7 @@
 import React, { Suspense, useContext, useState } from "react"
-import Queue from "p-queue"
+import Queue from "better-queue"
 import lodash from "lodash"
+import mitt from "mitt"
 
 import resources from "../resources"
 
@@ -15,14 +16,9 @@ import { useRecipeStep } from "./step-component"
 import { InputProvider } from "./input-provider"
 import { ResourceProvider, useResourceContext } from "./resource-provider"
 
-const queue = new Queue({ concurrency: 5, autoStart: false })
-
-const resultCache = new Map()
-const inFlightCache = new Map()
-
-const ModeContext = React.createContext({})
-const useMode = () => useContext(ModeContext)
-const ModeProvider = ModeContext.Provider
+const GlobalsContext = React.createContext({})
+const useGlobals = () => useContext(GlobalsContext)
+const GlobalsProvider = GlobalsContext.Provider
 
 const getUserProps = props => {
   // eslint-disable-next-line
@@ -34,14 +30,28 @@ const SetResourcesProvider = React.createContext()
 
 let resourcesCache
 
-const Wrapper = ({ children, inputs, isApply }) => {
+const Wrapper = ({
+  children,
+  inputs,
+  isApply,
+  resultCache,
+  inFlightCache,
+  queue,
+}) => {
   // eslint-disable-next-line
   const [resourcesList, setResources] = useState(resourcesCache || [])
   resourcesCache = resourcesList
 
   return (
     <ErrorBoundary>
-      <ModeProvider value={{ mode: isApply ? `apply` : `plan` }}>
+      <GlobalsProvider
+        value={{
+          mode: isApply ? `apply` : `plan`,
+          resultCache,
+          inFlightCache,
+          queue,
+        }}
+      >
         <SetResourcesProvider.Provider value={setResources}>
           <ResourceProvider value={resourcesList}>
             <InputProvider value={inputs}>
@@ -51,7 +61,7 @@ const Wrapper = ({ children, inputs, isApply }) => {
             </InputProvider>
           </ResourceProvider>
         </SetResourcesProvider.Provider>
-      </ModeProvider>
+      </GlobalsProvider>
     </ErrorBoundary>
   )
 }
@@ -63,7 +73,7 @@ const ResourceComponent = ({
   children,
   ...props
 }) => {
-  const { mode } = useMode()
+  const { mode, resultCache, inFlightCache, queue } = useGlobals()
   const step = useRecipeStep()
   const parentResourceContext = useParentResourceContext()
 
@@ -75,6 +85,9 @@ const ResourceComponent = ({
       root: process.cwd(),
       _uuid,
       mode,
+      resultCache,
+      inFlightCache,
+      queue,
     },
     props
   )
@@ -106,7 +119,7 @@ const validateResource = (resourceName, context, props) => {
 
 const handleResource = (resourceName, context, props) => {
   // Initialize
-  const { mode } = context
+  const { mode, resultCache, inFlightCache, queue } = context
 
   const trueKey = props._key ? props._key : context._uuid
 
@@ -173,22 +186,21 @@ const handleResource = (resourceName, context, props) => {
         resolve(cachedValue)
         updateResource(cachedValue)
       } else {
-        // if (fn == `create`) {
-        // console.log(`start creating resource:`, { context, props })
-        // }
         resources[resourceName][fn](context, props)
           .then(result => {
-            // if (fn == `create`) {
-            // console.log(`finish creating resource:`, { props })
-            // }
+            if (fn === `create`) {
+              result.isDone = true
+            }
             updateResource(result)
             inFlightCache.set(cacheKey, false)
             return result
           })
-          .then(result => resultCache.set(cacheKey, result))
+          .then(result => {
+            resultCache.set(cacheKey, result)
+            return result
+          })
           .then(resolve)
           .catch(e => {
-            console.log(e)
             if (e.name === `MissingInfoError`) {
               inFlightCache.delete(cacheKey)
             }
@@ -202,43 +214,110 @@ const handleResource = (resourceName, context, props) => {
 
   inFlightCache.set(cacheKey, promise)
 
-  queue.add(() => promise)
+  queue.push(promise)
 
   throw promise
 }
 
-const render = async (recipe, cb, inputs = {}, isApply) => {
+const render = (recipe, cb, inputs = {}, isApply, isStream, name) => {
+  const emitter = mitt()
   const plan = {}
 
+  const queue = new Queue(
+    async (job, cb) => {
+      const result = await job
+      cb(null, result)
+    },
+    { concurrent: 5 }
+  )
+
+  const resultCache = new Map()
+  const inFlightCache = new Map()
+
+  let result
+
   const recipeWithWrapper = (
-    <Wrapper inputs={inputs} isApply={isApply}>
+    <Wrapper
+      inputs={inputs}
+      isApply={isApply}
+      resultCache={resultCache}
+      inFlightCache={inFlightCache}
+      queue={queue}
+    >
       {recipe}
     </Wrapper>
   )
 
-  const renderResources = async () => {
-    queue.pause()
+  // Keep calling render until there's remaining resources to render.
+  // This let's resources that depend on other resources to pause until one finishes.
+  const renderResources = isDrained => {
+    result = RecipesReconciler.render(recipeWithWrapper, plan, name)
 
-    RecipesReconciler.render(recipeWithWrapper, plan)
+    const resources = transformToPlan(result)
 
-    if (!queue.size) {
-      return undefined
+    const isDone = () => {
+      // Mostly for validation stage that checks that there's no resources
+      // in the initial step — this done condition says no resources were found
+      // and there's no inflight resource work (resources will be empty until the
+      // first resource returns).
+      //
+      // We use "inFlightCache" because the queue doesn't immediately show up
+      // as having things in it.
+      if (resources.length === 0 && ![...inFlightCache.values()].some(a => a)) {
+        return true
+        // If there's still nothing on the queue and we've drained the queue, that means we're done.
+      } else if (isDrained && queue.length === 0) {
+        return true
+        // If there's one resource & it fails validation, it doesn't go into the queue
+        // so we check if inFlightCache is empty & all resources have an error.
+      } else if (
+        !resources.some(r => !r.error) &&
+        ![...inFlightCache.values()].some(a => a)
+      ) {
+        return true
+      }
+
+      return false
     }
 
-    queue.start()
-    await queue.onIdle()
-    return await renderResources()
+    if (isDone()) {
+      // Rerender with the resources and resolve the data from the cache.
+      emitter.emit(`done`, resources)
+    }
   }
 
-  try {
-    // Begin the "render loop" until there are no more resources being queued.
-    await renderResources()
+  const throttledRenderResources = lodash.throttle(renderResources, 30, {
+    trailing: false,
+  })
 
-    // Rerender with the resources and resolve the data from the cache
-    const result = RecipesReconciler.render(recipeWithWrapper, plan)
-    return transformToPlan(result)
-  } catch (e) {
-    throw e
+  queue.on(`task_finish`, function (taskId, r, stats) {
+    throttledRenderResources()
+
+    const resources = transformToPlan(result)
+    emitter.emit(`update`, resources)
+  })
+
+  queue.on(`drain`, () => {
+    renderResources(true)
+  })
+
+  // When there's no resources, renderResources finishes synchronously
+  // so wait for the next tick so the emitter listners can be setup first.
+  process.nextTick(() => renderResources())
+
+  if (isStream) {
+    return emitter
+  } else {
+    return new Promise((resolve, reject) => {
+      emitter.on(`*`, (type, e) => {
+        if (type === `done`) {
+          resolve(e)
+        }
+        if (type === `error`) {
+          reject(e)
+        }
+      })
+    })
   }
 }